+++ index.js
@@ -7,11 +7,6 @@ require(`./src/namespace`).populate();
 
 require(`./src/ajax`).applyProtocolWorkaround();
 
-// Inject CSS file
-require(`./src/css`).injectCss();
-
-HHM.ui = require('./src/ui/index');
-
 // Create plugin manager
 HHM.manager = new HHM.classes.PluginManager();
 
@@ -20,12 +15,6 @@ HHM.deferreds.roomLink = new $.Deferred();
 
 // Provides the config, waits for captcha solution and starts the plugin
 // manager
-let room = {};
-require(`./src/ui/config`).provideConfig()
-  .then(() => room = HHM.manager.provideRoom())
-  .then(() => HHM.manager.start(room))
-  .then(() => HHM.deferreds.managerStarted.promise())
-  .then(() => HHM.config.dryRun ? true : HHM.deferreds.roomLink.promise())
-  .then(() => HHM.ui.setHhmConfigAndIframeVisibility(true))
-  .then(() => HHM.config.trueHeadless ? true
-    : require(`./src/ui/plugins`).initialize());
+let room = HHM.manager.provideRoom();
+
+HHM.manager.start(room);